requirements.txt
langchain 
langgraph 
streamlit 
faiss-cpu 
chromadb 
python-dotenv
requests
langchain-ollama
langchain-ollama
